@@ -2,7 +2,12 @@
 
 namespace App\Providers;
 
+use App\Models\Company;
+use App\Models\Activity;
+use App\Policies\CompanyUserPolicy;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\ServiceProvider;
+use App\Policies\CompanyActivityPolicy;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -19,6 +24,7 @@ class AppServiceProvider extends ServiceProvider
      */
     public function boot(): void
     {
-        //
+        Gate::policy(Company::class, CompanyUserPolicy::class);
+        Gate::policy(Activity::class, CompanyActivityPolicy::class);
     }
 }
