@@ -9,9 +9,9 @@
         <div class="mx-auto max-w-7xl sm:px-6 lg:px-8">
             <div class="overflow-hidden bg-white shadow-sm sm:rounded-lg">
                 <div class="p-6 text-gray-900">
-                    @if(session('success'))
+                    @session('success'))
                         <div class="mb-6 bg-indigo-100 p-4 text-indigo-700">{{ session('success') }}</div>
-                    @endif
+                    @endsession
 
                     <div class="grid grid-cols-4 gap-5">
                         @forelse($activities as $activity)
