@@ -32,7 +32,7 @@ class RegistrationTest extends TestCase
         ]);
 
         $this->assertAuthenticated();
-        $response->assertRedirect(RouteServiceProvider::HOME);
+        $response->assertRedirect(route('home', absolute: false));
     }
 
     public function test_user_can_register_with_token_for_company_owner_role()
@@ -64,7 +64,7 @@ class RegistrationTest extends TestCase
 
         $this->assertAuthenticated();
 
-        $response->assertRedirect(RouteServiceProvider::HOME);
+        $response->assertRedirect(route('home', absolute: false));
     }
 
     public function test_user_can_register_with_token_for_guide_role()
@@ -96,6 +96,6 @@ class RegistrationTest extends TestCase
 
         $this->assertAuthenticated();
 
-        $response->assertRedirect(RouteServiceProvider::HOME);
+        $response->assertRedirect(route('home', absolute: false));
     }
 }
